requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor>=3.6.0
pymongo>=4.9.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
websockets==12.0
python-multipart==0.0.6
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pyjwt[crypto]==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0